the B-tree is just for integer

+++ unistd.h
@@ -9,17 +9,17 @@ struct B_Tree
 {
 	// point to an array, whose length is depended on the order that
 	// is decided by input.
-    float * p_key;
+    int * p_key;
 
     // point to a pointer array. there should be (key_counter + 1)
     // elements in the array.
     struct B_Tree * p_children_tree;
 
-    // to use this variable to judge
+    // to use this variable to judge if it is necessary to seperate.
     int key_counter;   
 };
 
 int ErrNum;  // A globle variable represent all kinds of errors.
 
-int insert(B_Tree * p_tree, float value);
-int search(B_Tree * p_tree, float value);
+int insert(B_Tree * p_tree, int value);
+int search(B_Tree * p_tree, int value);
